+++ src/clj/spring_lobby.clj
@@ -530,7 +530,6 @@
             by-server)))))
 
 
-
 (defn server-auto-resources [_old-state new-state old-server new-server]
   (when (:auto-get-resources new-state)
     (try
@@ -842,13 +841,11 @@
             new-replay (or (get parsed-replays-by-path new-replay-path)
                            (get online-bar-replays selected-replay-id))
 
-            old-game (-> old-replay :body :script-data :game)
-            old-mod (:gametype old-game)
-            old-map (:mapname old-game)
+            old-mod (:replay-mod-name old-replay)
+            old-map (:replay-map-name old-replay)
 
-            new-game (-> new-replay :body :script-data :game)
-            new-mod (:gametype new-game)
-            new-map (:mapname new-game)
+            new-mod (:replay-mod-name new-replay)
+            new-map (:replay-map-name new-replay)
 
             spring-root-path (fs/canonical-path spring-isolation-dir)
 
@@ -1018,7 +1015,7 @@
             (when (and selected-replay-id (not selected-replay-file))
               (when-let [local (->> parsed-replays-by-path
                                     vals
-                                    (filter (comp #{selected-replay-id} :game-id :header))
+                                    (filter (comp #{selected-replay-id} :replay-id))
                                     first
                                     :file)]
                 (log/info "Fixing selected replay from" selected-replay-id "to" local)
@@ -1144,17 +1141,17 @@
                              (filter
                                (fn [replay]
                                  (if filter-replay-min-players
-                                   (<= filter-replay-min-players (fx.replay/replay-player-count replay))
+                                   (<= filter-replay-min-players (:replay-player-count replay))
                                    true)))
                              (filter
                                (fn [replay]
                                  (if filter-replay-max-players
-                                   (<= (fx.replay/replay-player-count replay) filter-replay-max-players)
+                                   (<= (:replay-player-count replay) filter-replay-max-players)
                                    true)))
                              (filter
                                (fn [replay]
                                  (if filter-replay-min-skill
-                                   (if-let [avg (fx.replay/average-skill (fx.replay/replay-skills replay))]
+                                   (if-let [avg (:replay-average-skill replay)]
                                      (<= filter-replay-min-skill avg)
                                      false)
                                    true)))
@@ -1164,30 +1161,25 @@
                                    true
                                    (every?
                                      (some-fn
-                                       (partial includes-term? (fx.replay/replay-id replay))
-                                       (partial includes-term? (get replays-tags (fx.replay/replay-id replay)))
-                                       (partial includes-term? (-> replay :header :engine-version))
-                                       (partial includes-term? (-> replay :body :script-data :game :gametype))
-                                       (partial includes-term? (-> replay :body :script-data :game :mapname))
+                                       (partial includes-term? (:replay-id replay))
+                                       (partial includes-term? (get replays-tags (:replay-id replay)))
+                                       (partial includes-term? (:replay-engine-version replay))
+                                       (partial includes-term? (:replay-mod-name replay))
+                                       (partial includes-term? (:replay-map-name replay))
                                        (fn [term]
-                                         (let [players (some->> replay :body :script-data :game
-                                                                (filter (comp #(string/starts-with? % "player") name first))
-                                                                (filter
-                                                                  (if replays-filter-specs
-                                                                    (constantly true)
-                                                                    (some-fn
-                                                                      (comp #{0 "0"} :spectator second)
-                                                                      (comp not #(contains? % :spectator) second))))
-                                                                (map (comp #(some % [:name :username]) second))
-                                                                (map fx.replay/sanitize-replay-filter))]
-                                           (some #(includes-term? % term) players))))
+                                         (let [players (concat
+                                                         (mapcat identity (:replay-allyteam-player-names replay))
+                                                         (when replays-filter-specs
+                                                           (:replay-spec-names replay)))]
+                                           (some #(includes-term? % term)
+                                                 (map fx.replay/sanitize-replay-filter players)))))
                                      filter-terms))))
                              doall)
                 replays (if replays-window-dedupe
                           (:replays
                             (reduce ; dedupe by id
                               (fn [agg curr]
-                                (let [id (fx.replay/replay-id curr)]
+                                (let [id (:replay-id curr)]
                                   (cond
                                     (not id)
                                     (update agg :replays conj curr)
@@ -1200,7 +1192,7 @@
                                :seen-ids #{}}
                               replays))
                           replays)]
-            (swap! *state assoc :filtered-replays replays)
+            (swap! *state assoc :filtered-replays (doall replays))
             (log/info "Filtered replays in" (- (u/curr-millis) before) "ms"))
           (catch Exception e
             (log/error e "Error in :filter-replays state watcher")))
@@ -1351,13 +1343,26 @@
      (fn [] (.close chimer)))))
 
 
-(defn- valid-replay-fn [all-paths-set]
+(defn- old-valid-replay-fn [all-paths-set]
   (fn [[path replay]]
     (and
       (contains? all-paths-set path) ; remove missing files
       (-> replay :header :game-id) ; re-parse if no game id
       (not (-> replay :file-size zero?)) ; remove empty files
       (not (-> replay :game-type #{:invalid}))))) ; remove invalid
+
+(defn- valid-replay-fn [all-paths-set]
+  (fn [[path replay]]
+    (and
+      (contains? all-paths-set path) ; remove missing files
+      (:replay-id replay) ; re-parse if no replay id
+      (not (-> replay :file-size zero?)) ; remove empty files
+      (not (-> replay :game-type #{:invalid}))))) ; remove invalid
+
+(defn migrate-replay [replay]
+  (merge
+    (select-keys replay [:file :filename :file-size :source-name])
+    (replay/replay-metadata replay)))
 
 (defn- refresh-replays
   [state-atom]
@@ -1387,7 +1392,8 @@
                                    {:source-name source})]))
                             doall)]
     (log/info "Parsed" (count this-round) "of" (count todo) "new replays in" (- (u/curr-millis) before) "ms")
-    (let [valid-replay? (valid-replay-fn all-paths)
+    (let [old-valid-replay? (old-valid-replay-fn all-paths)
+          valid-replay? (valid-replay-fn all-paths)
           new-state (swap! state-atom
                       (fn [state]
                         (let [old-replays (:parsed-replays-by-path state)
@@ -1396,15 +1402,24 @@
                               valid-replays (->> all-replays
                                                  (filter valid-replay?)
                                                  (into {}))
-                              valid-replay-paths (set (keys valid-replays))
+                              migratable-replays (->> all-replays
+                                                      (remove valid-replay?)
+                                                      (filter old-valid-replay?))
+                              _ (log/info "Migrating" (count migratable-replays) "replays")
+                              migrated-replays (->> migratable-replays
+                                                    (map (fn [[path replay]] [path (migrate-replay replay)]))
+                                                    (into {}))
+                              _ (log/info "Migrated" (count migratable-replays) "replays")
+                              replays-by-path (merge valid-replays migrated-replays)
+                              valid-replay-paths (set (concat (keys replays-by-path)))
                               invalid-replay-paths (->> all-replays
-                                                        (remove valid-replay?)
+                                                        (remove (some-fn old-valid-replay? valid-replay?))
                                                         keys
                                                         (concat (:invalid-replay-paths state))
                                                         (remove valid-replay-paths)
                                                         set)]
                           (assoc state
-                                 :parsed-replays-by-path valid-replays
+                                 :parsed-replays-by-path replays-by-path
                                  :invalid-replay-paths invalid-replay-paths))))
           invalid-replay-paths (set (:invalid-replay-paths new-state))
           valid-next-round (remove
@@ -1937,7 +1952,7 @@
   (let [chimer
         (chime/chime-at
           (chime/periodic-seq
-            (java-time/plus (java-time/instant) (java-time/duration 30 :seconds))
+            (java-time/plus (java-time/instant))
             (java-time/duration 30 :seconds))
           (fn [_chimestamp]
             (log/info "Updating music queue")
@@ -2390,7 +2405,7 @@
       (if (and replay-file (fs/exists? replay-file))
         (do
           (log/info "Updating replay details for" replay-file)
-          (let [replay-details (replay/parse-replay replay-file {:parse-stream true})]
+          (let [replay-details (replay/parse-replay replay-file {:details true :parse-stream true})]
             (log/info "Got replay details for" replay-file (keys replay-details))
             (swap! *state update :replay-details cache/miss cache-key replay-details)))
         (do
@@ -3763,7 +3778,7 @@
 
 
 (defmethod task-handler ::update-rapid
-  [{:keys [engine-version mod-name rapid-repo spring-isolation-dir] :as e}]
+  [{:keys [engine-version mod-name rapid-id rapid-repo spring-isolation-dir] :as e}]
   (swap! *state assoc :rapid-update true)
   (let [before (u/curr-millis)
         {:keys [by-spring-root file-cache] :as state} @*state ; TODO remove deref
@@ -3775,22 +3790,18 @@
                          preferred-engine-details
                          (->> engines
                               (filter (comp fs/canonical-path :file))
-                              first))]
+                              first))
+        rapid-repo (or rapid-repo
+                       (resource/mod-repo-name mod-name))
+        rapid-id (or rapid-id
+                     (str rapid-repo ":test"))]
     (if (and engine-details (:file engine-details))
       (do
         (log/info "Initializing rapid by calling download")
         (deref
           (event-handler
             {:event/type ::rapid-download
-             :rapid-id
-             (or (when rapid-repo
-                   (str rapid-repo ":test"))
-                 (when mod-name
-                   (cond
-                     (string/includes? mod-name "Beyond All Reason") "byar:test"
-                     :else nil))
-                 "i18n:test")
-             ; TODO how else to init rapid without download...
+             :rapid-id rapid-id
              :engine-file (:file engine-details)
              :spring-isolation-dir spring-isolation-dir})))
       (log/warn "No engine details to do rapid init"))
@@ -3816,8 +3827,8 @@
                               reverse)
           _ (log/info "Found" (count rapid-versions) "rapid versions")
           rapid-data-by-hash (->> rapid-versions
-                              (map (juxt :hash identity))
-                              (into {}))
+                                  (map (juxt :hash identity))
+                                  (into {}))
           rapid-data-by-id (->> rapid-versions
                                 (map (juxt :id identity))
                                 (into {}))
@@ -3833,7 +3844,13 @@
               (update :rapid-data-by-id merge rapid-data-by-id)
               (update :rapid-data-by-version merge rapid-data-by-version)
               (update :rapid-versions (fn [old-versions]
-                                        (set (concat old-versions rapid-versions))))
+                                        (set
+                                          (vals
+                                            (merge
+                                              (into {}
+                                                (map (juxt :id identity) old-versions))
+                                              (into {}
+                                                (map (juxt :id identity) rapid-versions)))))))
               (update :file-cache merge new-files))))
       (log/info "Updated rapid repo data in" (- (u/curr-millis) before) "ms"))
     (add-task! *state
@@ -4338,13 +4355,17 @@
                         (map (fn [{:keys [playerId] :as spec}]
                                [(str "player" playerId)
                                 {:username (:name spec)
-                                 :spectator 1}])))]
-    (-> replay
-        (assoc :source-name "BAR Online")
-        (assoc :body {:script-data {:game (into {} (concat (:hostSettings replay) players teams spectators))}})
-        (assoc :header {:unix-time (quot (inst-ms (java-time/instant (:startTime replay))) 1000)})
-        (assoc :player-counts player-counts)
-        (assoc :game-type (u/game-type player-counts)))))
+                                 :spectator 1}])))
+        r
+        (-> replay
+            (assoc :source-name "BAR Online")
+            (assoc :body {:script-data {:game (into {} (concat (:hostSettings replay) players teams spectators))}})
+            (assoc :header {:unix-time (quot (inst-ms (java-time/instant (:startTime replay))) 1000)})
+            (assoc :player-counts player-counts)
+            (assoc :game-type (u/game-type player-counts)))]
+    (merge
+      r
+      {:replay-map-name (-> r :Map :scriptName)})))
 
 (defmethod task-handler ::download-bar-replays [{:keys [page]}]
   (let [new-bar-replays (->> (http/get-bar-replays {:page page})
@@ -4353,7 +4374,7 @@
     (let [[before after] (swap-vals! *state
                            (fn [state]
                              (-> state
-                                 (assoc :bar-replays-page ((fnil inc 0) (int (u/to-number page))))
+                                 (assoc :bar-replays-page ((fnil inc 0) ((fnil int 0) (u/to-number page))))
                                  (update :online-bar-replays
                                    (fn [online-bar-replays]
                                      (u/deep-merge
@@ -4578,13 +4599,13 @@
 
 (def state-watch-chimers
   [
-   [:auto-get-resources-watcher auto-get-resources-watcher]
-   [:battle-map-details-watcher battle-map-details-watcher]
-   [:battle-mod-details-watcher battle-mod-details-watcher 15]
+   [:auto-get-resources-watcher auto-get-resources-watcher 2]
+   [:battle-map-details-watcher battle-map-details-watcher 2]
+   [:battle-mod-details-watcher battle-mod-details-watcher 2]
    [:fix-spring-isolation-dir-watcher fix-spring-isolation-dir-watcher 10]
    [:replay-map-and-mod-details-watcher replay-map-and-mod-details-watcher]
    [:spring-isolation-dir-changed-watcher spring-isolation-dir-changed-watcher 10]
-   [:update-battle-status-sync-watcher update-battle-status-sync-watcher 15]])
+   [:update-battle-status-sync-watcher update-battle-status-sync-watcher 2]])
 
 
 (defn ipc-handler
@@ -4676,8 +4697,8 @@
            (add-task! state-atom {::task-type ::refresh-mods})
            (async/<!! (async/timeout wait-init-tasks-ms))
            (add-task! state-atom {::task-type ::refresh-maps})
-           (async/<!! (async/timeout wait-init-tasks-ms))
-           (add-task! state-atom {::task-type ::refresh-replays})
+           ;(async/<!! (async/timeout wait-init-tasks-ms))
+           ;(add-task! state-atom {::task-type ::refresh-replays})
            (async/<!! (async/timeout wait-init-tasks-ms))
            (add-task! state-atom {::task-type ::update-rapid})
            (async/<!! (async/timeout wait-init-tasks-ms))